fix(mcp-catalog): preserve registry trust boundaries
- support transparent proxy DNS without allowing direct private targets
- separate registry wire data from catalog normalization
- reject malformed registry payloads at the API boundary
